Do not use context to configure ADR command, use an execute function with all the parameters instead
